VIH-8628 Fix the Security token exception arising from the GetVideoControlStatuses end-point
