+++ VideoWeb/VideoWeb/Controllers/ConferenceManagementController.cs
@@ -85,75 +85,6 @@ namespace VideoWeb.Controllers
         }
                 
         /// <summary>
-        /// Updates the video control statuses for the conference
-        /// </summary>
-        /// <param name="conferenceId">conference id</param>
-        /// <returns>Ok status</returns>
-        /// <returns>Forbidden status</returns>
-        /// <returns>Not Found status</returns>
-        [HttpPut("{conferenceId}/setVideoControlStatuses")]
-        [SwaggerOperation(OperationId = "SetVideoControlStatusesForConference")]
-        [ProducesResponseType((int)HttpStatusCode.Accepted)]
-        [ProducesResponseType((int)HttpStatusCode.Forbidden)]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        public async Task<IActionResult> SetVideoControlStatusesForConference(Guid conferenceId, [FromBody]SetConferenceVideoControlStatusesRequest setVideoControlStatusesRequest)
-        {
-            try
-            {
-                var mapper = _mapperFactory.Get<SetConferenceVideoControlStatusesRequest, ConferenceVideoControlStatuses>();
-                var videoControlStatuses = mapper.Map(setVideoControlStatusesRequest);
-                
-                _logger.LogDebug("Setting the video control statuses for {conferenceId}", conferenceId);
-                await _conferenceVideoControlStatusService.SetVideoControlStateForConference(conferenceId, videoControlStatuses);
-
-                _logger.LogTrace("Set video control statuses ({videoControlStatuses}) for {conferenceId}", videoControlStatuses, conferenceId);
-                return Accepted();
-            }
-            catch (Exception exception)
-            {
-                _logger.LogError(exception, "Could not set video control statuses for {conferenceId} an unkown exception was thrown", conferenceId);
-                return StatusCode((int)HttpStatusCode.InternalServerError);
-            }
-        }
-        
-        /// <summary>
-        /// Returns the video control statuses for the conference
-        /// </summary>
-        /// <param name="conferenceId">conference id</param>
-        /// <returns>Ok status</returns>
-        /// <returns>Forbidden status</returns>
-        /// <returns>Not Found status</returns>
-        [HttpGet("{conferenceId}/getVideoControlStatuses")]
-        [SwaggerOperation(OperationId = "GetVideoControlStatusesForConference")]
-        [ProducesResponseType(typeof(ConferenceVideoControlStatuses), (int)HttpStatusCode.OK)]
-        [ProducesResponseType((int)HttpStatusCode.Forbidden)]
-        [ProducesResponseType(typeof(ConferenceVideoControlStatuses), (int)HttpStatusCode.NotFound)]
-        public async Task<IActionResult> GetVideoControlStatusesForConference(Guid conferenceId)
-        {
-            try
-            {
-                _logger.LogDebug("Getting the video control statuses for {conferenceId}", conferenceId);
-                var videoControlStatuses = await _conferenceVideoControlStatusService.GetVideoControlStateForConference(conferenceId);
-
-                if (videoControlStatuses == null) {
-                    _logger.LogWarning("video control statuses didn't have a value returning NotFound. This was for {conferenceId}", conferenceId);
-                    return NotFound(new ConferenceVideoControlStatuses()
-                    {
-                        ParticipantIdToVideoControlStatusMap = new Dictionary<string, VideoControlStatus>()
-                    });
-                }
-
-                _logger.LogTrace("Got video control statuses ({videoControlStatuses}) for {conferenceId}", videoControlStatuses, conferenceId);
-                return Ok(videoControlStatuses);
-            }
-            catch (Exception exception)
-            {
-                _logger.LogError(exception, "Could not get video control statuses for {conferenceId} an unkown exception was thrown", conferenceId);
-                return StatusCode((int)HttpStatusCode.InternalServerError);
-            }
-        }
-
-        /// <summary>
         /// Returns the active layout for a conference
         /// </summary>
         /// <param name="conferenceId">conference id</param>
